+++ src/main/java/igentuman/nc/radiation/data/PlayerRadiation.java
@@ -23,7 +23,7 @@ public class PlayerRadiation implements IPlayerRadiationCapability {
     private final double decaySpeed = RADIATION_CONFIG.DECAY_SPEED_FOR_PLAYER.get();
 
     public Level level;
-    private int radiation = 0;
+    private long radiation = 0;
     private int timestamp = 0;
 
     private int contaminationStage = 0;
@@ -42,7 +42,7 @@ public class PlayerRadiation implements IPlayerRadiationCapability {
     @Override
     public CompoundTag serializeNBT() {
         CompoundTag tag = new CompoundTag();
-        tag.putInt("radiation", radiation);
+        tag.putLong("radiation", radiation);
         tag.putInt("timestamp", timestamp);
         return tag;
     }
@@ -50,7 +50,7 @@ public class PlayerRadiation implements IPlayerRadiationCapability {
     @Override
     public void deserializeNBT(CompoundTag nbt) {
         CompoundTag radiationTag = nbt.getCompound("radiation");
-        radiation = radiationTag.getInt("radiation");
+        radiation = radiationTag.getLong("radiation");
         timestamp = radiationTag.getInt("timestamp");
     }
 
@@ -172,12 +172,12 @@ public class PlayerRadiation implements IPlayerRadiationCapability {
     }
 
     @Override
-    public int getRadiation() {
+    public long getRadiation() {
         return radiation;
     }
 
     @Override
-    public void setRadiation(int radiation) {
+    public void setRadiation(long radiation) {
         this.radiation = radiation;
     }
